+++ src/app/(auth)/Signin/page.tsx
@@ -1,4 +1,6 @@
+import {
 
+} from "@nextui-org/react"
 
 function Signin() {
     return (